Use try_cast instead of cast
Addresses issue https://github.com/tidyverse/dbplyr/issues/380
Preferable for MSSQL 2012 and above, as failure to CAST results in an error. By contrast TRY_CAST (mostly) returns a NA in circumstances where the cast fails. However, this is a MSSQL (2012+) feature, and could break when used with other database backends.